updated go-sdk to latest version and required changes
