+++ providers/flagd/go.mod
@@ -4,18 +4,23 @@ go 1.18
 
 require (
 	github.com/open-feature/flagd v0.0.4
-	github.com/open-feature/golang-sdk v0.0.0-20220712142818-221a71bb32b8
+	github.com/open-feature/golang-sdk v0.0.0-20220729134802-91d68f3ffa9f
 	github.com/sirupsen/logrus v1.9.0
+	github.com/stretchr/testify v1.7.4
 	go.buf.build/grpc/go/open-feature/flagd v1.3.1
 	google.golang.org/grpc v1.48.0
 	google.golang.org/protobuf v1.28.0
 )
 
 require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/google/go-cmp v0.5.8 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	golang.org/x/net v0.0.0-20220617184016-355a448f1bc9 // indirect
 	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
 	golang.org/x/text v0.3.7 // indirect
 	google.golang.org/genproto v0.0.0-20220714211235-042d03aeabc9 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
