+++ providers/flagd/pkg/service/grpc/service-resolve-boolean_test.go
@@ -1,119 +1,131 @@
-package tests
+package grpc_service
 
 import (
 	"errors"
 	"testing"
 
 	models "github.com/open-feature/flagd/pkg/model"
-	service "github.com/open-feature/golang-sdk-contrib/providers/flagd/pkg/service/grpc"
-	"github.com/open-feature/golang-sdk-contrib/providers/flagd/pkg/service/grpc/tests/mocks"
+	"github.com/open-feature/golang-sdk-contrib/providers/flagd/pkg/service/grpc/mocks"
+	of "github.com/open-feature/golang-sdk/pkg/openfeature"
 	"github.com/stretchr/testify/assert"
 	schemaV1 "go.buf.build/grpc/go/open-feature/flagd/schema/v1"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
 
-type TestServiceResolveStringArgs struct {
+type TestServiceResolveBooleanArgs struct {
 	name string
-	mocks.MockResolveStringArgs
+	mocks.MockResolveBooleanArgs
 	nilClient bool
 
 	flagKey string
-	evCtx   interface{}
+	evCtx   of.EvaluationContext
 
-	value     string
+	value     bool
 	variant   string
 	reason    string
 	err       error
 	customErr string
 }
 
-func TestServiceResolveString(t *testing.T) {
-	tests := []TestServiceResolveStringArgs{
+func TestServiceResolveBoolean(t *testing.T) {
+	tests := []TestServiceResolveBooleanArgs{
 		{
 			name: "happy path",
-			MockResolveStringArgs: mocks.MockResolveStringArgs{
+			MockResolveBooleanArgs: mocks.MockResolveBooleanArgs{
 				InFK: "bool",
-				InCtx: map[string]interface{}{
-					"dog": "cat",
+				InCtx: of.EvaluationContext{
+					Attributes: map[string]interface{}{
+						"con": "text",
+					},
 				},
-				Out: &schemaV1.ResolveStringResponse{
-					Value:   "on",
+				Out: &schemaV1.ResolveBooleanResponse{
+					Value:   true,
 					Variant: "on",
 					Reason:  models.StaticReason,
 				},
 			},
 			flagKey: "bool",
-			evCtx: map[string]interface{}{
-				"dog": "cat",
+			evCtx: of.EvaluationContext{
+				Attributes: map[string]interface{}{
+					"con": "text",
+				},
 			},
+			value:   true,
 			variant: "on",
-			value:   "on",
 			reason:  models.StaticReason,
 			err:     nil,
 		},
 		{
-			name:    "FormatAsStructpb fails",
-			flagKey: "bool",
-			evCtx:   "not a map[string]interface{}!",
-			reason:  models.ErrorReason,
-			err:     errors.New(models.ParseErrorCode),
-		},
-		{
 			name: "custom error response",
-			MockResolveStringArgs: mocks.MockResolveStringArgs{
+			MockResolveBooleanArgs: mocks.MockResolveBooleanArgs{
 				InFK: "bool",
-				InCtx: map[string]interface{}{
-					"dog": "cat",
+				InCtx: of.EvaluationContext{
+					Attributes: map[string]interface{}{
+						"con": "text",
+					},
 				},
 				OutErr: status.Error(codes.NotFound, "custom message"),
 			},
 			flagKey: "bool",
-			evCtx: map[string]interface{}{
-				"dog": "cat",
+			evCtx: of.EvaluationContext{
+				Attributes: map[string]interface{}{
+					"con": "text",
+				},
 			},
+			value:     false,
 			reason:    models.ErrorReason,
 			customErr: "CUSTOM ERROR",
 			err:       errors.New("CUSTOM ERROR"),
 		},
 		{
-			name: "nil client",
-			MockResolveStringArgs: mocks.MockResolveStringArgs{
-				InFK: "bool",
-				InCtx: map[string]interface{}{
-					"dog": "cat",
-				},
-				OutErr: status.Error(codes.NotFound, "custom message"),
-			},
-			flagKey: "bool",
-			evCtx: map[string]interface{}{
-				"dog": "cat",
-			},
-			nilClient: true,
-			reason:    models.ErrorReason,
-			err:       errors.New("CONNECTION_ERROR"),
-		},
-		{
 			name: "error parse failure",
-			MockResolveStringArgs: mocks.MockResolveStringArgs{
+			MockResolveBooleanArgs: mocks.MockResolveBooleanArgs{
 				InFK: "bool",
-				InCtx: map[string]interface{}{
-					"dog": "cat",
+				InCtx: of.EvaluationContext{
+					Attributes: map[string]interface{}{
+						"con": "text",
+					},
 				},
 				OutErr: status.Error(codes.NotFound, "custom message"),
 			},
 			flagKey: "bool",
-			evCtx: map[string]interface{}{
-				"dog": "cat",
+			evCtx: of.EvaluationContext{
+				Attributes: map[string]interface{}{
+					"con": "text",
+				},
 			},
+			value:  false,
 			reason: models.ErrorReason,
 			err:    errors.New("GENERAL"),
+		},
+		{
+			name: "nil client",
+			MockResolveBooleanArgs: mocks.MockResolveBooleanArgs{
+				InFK: "bool",
+				InCtx: of.EvaluationContext{
+					Attributes: map[string]interface{}{
+						"con": "text",
+					},
+				},
+				OutErr: status.Error(codes.NotFound, "custom message"),
+			},
+			flagKey: "bool",
+			evCtx: of.EvaluationContext{
+				Attributes: map[string]interface{}{
+					"con": "text",
+				},
+			},
+			nilClient: true,
+			value:     false,
+			reason:    models.ErrorReason,
+			err:       errors.New("CONNECTION_ERROR"),
 		},
 	}
 
 	for _, test := range tests {
 		if test.customErr != "" {
-			st, ok := status.FromError(test.MockResolveStringArgs.OutErr)
+			st, ok := status.FromError(test.MockResolveBooleanArgs.OutErr)
 			if !ok {
 				t.Errorf("%s: malformed error status recieved, cannot attach custom properties", test.name)
 			}
@@ -124,16 +136,16 @@ func TestServiceResolveString(t *testing.T) {
 			if err != nil {
 				t.Error(err)
 			}
-			test.MockResolveStringArgs.OutErr = stWD.Err()
+			test.MockResolveBooleanArgs.OutErr = stWD.Err()
 		}
-		srv := service.GRPCService{
+		srv := GRPCService{
 			Client: &mocks.MockClient{
 				ReturnNilClient: test.nilClient,
-				RSArgs:          test.MockResolveStringArgs,
+				RBArgs:          test.MockResolveBooleanArgs,
 				Testing:         t,
 			},
 		}
-		res, err := srv.ResolveString(test.flagKey, test.evCtx)
+		res, err := srv.ResolveBoolean(test.flagKey, test.evCtx)
 		if test.err != nil && !assert.EqualError(t, err, test.err.Error()) {
 			t.Errorf("%s: unexpected error received, expected %v, got %v", test.name, test.err, err)
 		}
