+++ providers/flagd/pkg/service/iservice.go
@@ -5,11 +5,9 @@ import (
 	schemaV1 "go.buf.build/grpc/go/open-feature/flagd/schema/v1"
 )
 
-type IServiceOption func(IService)
-
 type IService interface {
-	ResolveBoolean(string, of.EvaluationContext, ...IServiceOption) (*schemaV1.ResolveBooleanResponse, error)
-	ResolveString(string, of.EvaluationContext, ...IServiceOption) (*schemaV1.ResolveStringResponse, error)
-	ResolveNumber(string, of.EvaluationContext, ...IServiceOption) (*schemaV1.ResolveNumberResponse, error)
-	ResolveObject(string, of.EvaluationContext, ...IServiceOption) (*schemaV1.ResolveObjectResponse, error)
+	ResolveBoolean(string, of.EvaluationContext) (*schemaV1.ResolveBooleanResponse, error)
+	ResolveString(string, of.EvaluationContext) (*schemaV1.ResolveStringResponse, error)
+	ResolveNumber(string, of.EvaluationContext) (*schemaV1.ResolveNumberResponse, error)
+	ResolveObject(string, of.EvaluationContext) (*schemaV1.ResolveObjectResponse, error)
 }
